+++ sites/partners/src/listings/PaperListingForm/sections/ApplicationDates.tsx
@@ -3,8 +3,6 @@ import { useWatch, useFormContext } from "react-hook-form"
 import { YesNoAnswer } from "../../../applications/PaperApplicationForm/FormTypes"
 import { getDetailFieldDate, getDetailFieldTime } from "../../PaperListingDetails/sections/helpers"
 import dayjs from "dayjs"
-import utc from "dayjs/plugin/utc"
-dayjs.extend(utc)
 
 import {
   t,
@@ -44,7 +42,7 @@ const ApplicationDates = ({
 
   const openHouseTableData = useMemo(() => {
     return openHouseEvents.map((event) => {
-      const { startTime, endTime, url, tempId } = event
+      const { startTime, endTime, url } = event
 
       return {
         date: startTime && getDetailFieldDate(startTime),
@@ -70,7 +68,7 @@ const ApplicationDates = ({
             <Button
               type="button"
               className="font-semibold uppercase text-red-700"
-              onClick={() => setModalDeleteOpenHouse(tempId)}
+              onClick={() => setModalDeleteOpenHouse(event)}
               unstyled
             >
               {t("t.delete")}
@@ -92,23 +90,24 @@ const ApplicationDates = ({
   })
 
   const [drawerOpenHouse, setDrawerOpenHouse] = useState<TempEvent | boolean>(false)
-  const [modalDeleteOpenHouse, setModalDeleteOpenHouse] = useState<string | null>(null)
+  const [modalDeleteOpenHouse, setModalDeleteOpenHouse] = useState<TempEvent | null>(null)
 
   const onOpenHouseEventsSubmit = (event: TempEvent) => {
     setDrawerOpenHouse(false)
 
-    const eventsWithoutEdited = openHouseEvents.filter((item) => item.tempId !== event.tempId)
+    const eventsWithoutEdited = openHouseEvents.filter((item) => {
+      return event.id ? event.id !== item.id : event.tempId !== item.tempId
+    })
 
-    // determine if event is currently edited
-    if (eventsWithoutEdited.length !== openHouseEvents.length) {
-      setOpenHouseEvents([...eventsWithoutEdited, event])
-    } else {
-      setOpenHouseEvents([...openHouseEvents, event])
-    }
+    setOpenHouseEvents(
+      [...eventsWithoutEdited, event].sort((a, b) =>
+        dayjs(a.startTime).isAfter(b.startTime) ? 1 : -1
+      )
+    )
   }
 
-  const onOpenHouseEventDelete = (tempId: string) => {
-    const newEvents = openHouseEvents.filter((event) => event.tempId !== tempId)
+  const onOpenHouseEventDelete = (eventToDelete: TempEvent) => {
+    const newEvents = openHouseEvents.filter((event) => event !== eventToDelete)
     setOpenHouseEvents(newEvents)
     setModalDeleteOpenHouse(null)
   }
@@ -132,13 +131,13 @@ const ApplicationDates = ({
             disabled={enableDueDate === YesNoAnswer.No}
             defaultDate={{
               month: listing?.applicationDueDate
-                ? dayjs(new Date(listing?.applicationDueDate)).utc().format("MM")
+                ? dayjs(new Date(listing?.applicationDueDate)).format("MM")
                 : null,
               day: listing?.applicationDueDate
-                ? dayjs(new Date(listing?.applicationDueDate)).utc().format("DD")
+                ? dayjs(new Date(listing?.applicationDueDate)).format("DD")
                 : null,
               year: listing?.applicationDueDate
-                ? dayjs(new Date(listing?.applicationDueDate)).utc().format("YYYY")
+                ? dayjs(new Date(listing?.applicationDueDate)).format("YYYY")
                 : null,
             }}
           />
@@ -150,16 +149,26 @@ const ApplicationDates = ({
             watch={watch}
             disabled={enableDueDate === YesNoAnswer.No}
             defaultValues={{
-              hours: listing?.applicationDueTime
-                ? dayjs(new Date(listing?.applicationDueTime)).format("hh")
-                : null,
-              minutes: listing?.applicationDueTime
-                ? dayjs(new Date(listing?.applicationDueTime)).format("mm")
-                : null,
-              seconds: listing?.applicationDueTime
-                ? dayjs(new Date(listing?.applicationDueTime)).format("ss")
-                : null,
-              period: new Date(listing?.applicationDueTime).getHours() >= 12 ? "pm" : "am",
+              hours: listing?.applicationDueDate
+                ? dayjs(new Date(listing?.applicationDueDate)).format("hh")
+                : enableDueDate === YesNoAnswer.No
+                ? null
+                : "05",
+              minutes: listing?.applicationDueDate
+                ? dayjs(new Date(listing?.applicationDueDate)).format("mm")
+                : enableDueDate === YesNoAnswer.No
+                ? null
+                : "00",
+              seconds: listing?.applicationDueDate
+                ? dayjs(new Date(listing?.applicationDueDate)).format("ss")
+                : enableDueDate === YesNoAnswer.No
+                ? null
+                : "00",
+              period: listing?.applicationDueDate
+                ? new Date(listing?.applicationDueDate).getHours() >= 12
+                  ? "pm"
+                  : "am"
+                : "pm",
             }}
           />
         </GridSection>